refactor: client profile dashboard into sub components and directory tidying
