+++ frontend/src/components/clients/wizard/steps/MedicalProfileStep.tsx
