+++ frontend/src/firebase/googleCalendarService.js
@@ -187,7 +187,7 @@ function buildEventPayload(event) {
         ],
       };
     }
-  } catch (err) {
+  } catch {
     // If reminder conversion fails, do not break event creation/update — skip Google reminders.
     // Intentionally swallow errors here to avoid affecting the primary flow.
   }
